Fixed a few warnings

+++ DegreeProject/DegreeProject/Source/Game/Components/Map.h
@@ -225,9 +225,9 @@ struct Map : public Component
 		return *this;
 	}
 
-	size_t GetRegionPosition(const char& c)
+	int GetRegionPosition(const char& c)
 	{
-		for (size_t i = 0; i < m_Regions.size(); ++i)
+		for (int i = 0; i < m_Regions.size(); ++i)
 		{
 			if (m_Regions[i].m_MapChar == c)
 			{
